services/license-extraction-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.6
python-multipart==0.0.9
httpx==0.27.2
tenacity==9.0.0
langdetect==1.0.9
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10
python-json-logger==2.0.7
pytest==8.3.3
pytest-asyncio==0.24.0
